refactor server run and persist username in events table
